refactor(updates): align feature structure with devices pattern
Move main process code to dedicated main/ subdirectory and consolidate
orchestrators with services for consistency across features:

- Move update.manager.js to updates/main/ (matches devices/main/)
- Move update.orchestrator.js to updates/services/ (matches other features)
- Update all import paths in containers and tests
- Remove empty orchestrators/ directory

This aligns the updates feature with the established pattern where main
process code lives in a main/ subdirectory and renderer orchestrators
are colocated with services.
